[releases] add channel support for releases
so the version of "0.1.1-alpha@latest" will result in the version
"0.1.1-alpha" and the channel "latest" (otherwise our semversion sorting fails)

furthermore add (not yet used) getRelease() function which would take care of checking channels and a semantic version string.
Nvm this begins to go out of scope.. installLatest() is already what we need.